feat: add describe() method to Series
